Updated README.md
This project is a random password generator with a button. When you first open the file, you see a title, the text box where the password will be generated, and a button to generate the password. When you click the generate password button, a prompt immediately pops up asking specific questions about the passwords criteria. Once all of these questions are answered, a password will pop up in the text box depending on the criterias entered by the user.